backend/pyproject.toml
[project]
name = "contractlens"
version = "0.1.0"
description = "Analisador de contratos"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.128.7",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "bcrypt>=5.0.0",
    "python-jose>=3.5.0",
    "pydantic-settings>=2.12.0",
    "pydantic-extra-types>=2.11.0",
    "python-multipart>=0.0.22",
    "email-validator>=2.3.0",
    "PyYAML>=6.0.3",
    "Jinja2>=3.1.6",
    "aiosmtplib>=5.1.0",
    "aiofiles>=25.1.0",
    "websockets>=16.0"
]

[project.optional-dependencies]
dev = [
    "pytest-asyncio>=1.3.0",
    "testcontainers>=4.14.1"
]